+++ entry/tests/test_view.py
@@ -5178,3 +5178,27 @@ class ViewTest(AironeViewTest):
             Entry.objects.get(id=entry.id).attrs.count(),
             Entity.objects.get(id=entity.id).attrs.count(),
         )
+
+    def test_acl_is_not_editable_when_entity_has_not_full_permission(self):
+        user = self.guest_login()
+        for (index, acltype) in enumerate([ACLType.Readable, ACLType.Writable, ACLType.Nothing]):
+            entity = self.create_entity(
+                user,
+                "Test Another Entity %d" % index,
+                attrs=[{"name": "attr", "type": AttrTypeValue["string"]}],
+                is_public=False,
+                default_permission=acltype.id,
+            )
+            # check Entry's ACL couldn't editable
+            entry = self.add_entry(user, "TestEntry", entity)
+            resp = self.client.get(reverse("entry:acl", args=[entry.id]))
+            self.assertEqual(resp.status_code, 400)
+            self.assertEqual(resp.content.decode("utf-8"), "You don't have permission to access this object")
+
+            # check Attribute's and EntityAttr's ACL couldn't editable
+            attr = entry.attrs.last()
+            self.assertEqual(attr.schema.name, "attr")
+            for instance in [attr, attr.schema]:
+                resp = self.client.get(reverse("acl:index", args=[instance.id]))
+                self.assertEqual(resp.status_code, 400)
+                self.assertEqual(resp.content.decode("utf-8"), "You don't have permission to access this object")