+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 ### Changed
 
 ### Fixed
+* Fixed a bug to be able to edit Entry, Attribute and EntityAttr's ACL without Full permission
+  Contributed by @userlocalhost
 
 ## v3.43.0